Login page style revamp
