--- views/LayoutView.php
@@ -1,7 +1,0 @@
-<?php
-
-class LayoutView{
-    
-}
-
-?>
